export: Support generating bin files without ".bin" suffix
Currently "-ftb" generates files with the following extensions:

    ".img.bin", ".map.bin", ".meta.bin", ".pal.bin"

This is useful when adding the files as data to your binary because
common Makefiles automatically handle files with the ".bin" extension.

However, this is useless when the files are saved to a filesystem, such
as the NitroFS filesystem of NDS ROMs. It only makes file names longer.
For this reason, a new option has been added. Using "-ftB" will generate
the following extensions instead:

    ".img", ".map", ".meta", ".pal"