libgrit: Add transparent color to palette if not found
Previously, if the transparent color provided by the user (with
"-gTFF00FF", for example) wasn't found in the palette, whichever color
was used in index 0 would become the transparent color.

This is surprising for users. Imagine you have 100 sprites and you use
magenta as a common transparent color. Maybe one of the sprites doesn't
have any transparent pixel, but you will still load it with the same
code as the others, so color 0 will still be transparent.

This commit forces the palette to have the transparent color provided by
the user, and sets it to index 0 so that index 0 is never used by any
useful color.