libgrit: Export the right palette size for all formats
Before this patch all exported palettes had 256 colors in them, even if
the format was 1, 2 or 4 bpp (or one of the special texture formats of
the NDS).

This patch sets the right size so that the correct number of colors is
exported. However, this patch doesn't exclude colors that aren't used.