libgrit: Fail if there are too many tiles for the output format
In some cases it can be quite hard to notice that a map hasn't been
exported correctly. The current makefile of BlocksDS silences everything
but errors. If warnings are printed there can be too much text in the
build output, and developers will just ignore it.

This commit makes the export process fail if there are too many tiles in
the image after doing all the tile set reduction optimizations.